UI - Adding Functionality to Graph Buttons
In reference to issues #37, #20, Bev and I have added actual functionality to the ChartActivity view that pops up once you select a stock from the list. Users can now click on one of 3 buttons: 7 Day, 52wk, 12 Mths to change the time interval of the stock graph being displayed in the popup. Minor bugs to fix include a bouncing menu bar and the fact that you have to close the popup in order to see the changes take place.