Added Lib for RssFeed, Added Button and New Popup window
In Reference to issue #56, I've added a new Library called Simple-Rss2-Android as a dependency because it exists and it's convenient. I've added a button to the chart popup window, which triggers a callback to another popup window to open on top. As of now, window is empty but it's getting there.