Added functionality for viewing RSS
In reference to issue #56, this commit actually parses the response in XML to text values that can be seen in the application itself. So as it is right now, when clicking on View Newsfeed from within the Chart Popup, the user can see three elements per RSS story. Title, Date, Link. However, it is unformatted text and the links are untappable. These things are to be added next commit.